[Behat] Fix order show and thank you pages
