+++ src/Sylius/Behat/Resources/config/services/pages/shop.xml
@@ -15,6 +15,7 @@
     <imports>
         <import resource="shop/account.xml" />
         <import resource="shop/checkout.xml" />
+        <import resource="shop/order.xml" />
         <import resource="shop/product.xml" />
         <import resource="shop/static_content.xml" />
         <import resource="shop/taxon.xml" />
