+++ src/Sylius/Behat/Resources/config/services/pages/shop/order.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Paweł Jędrzejewski
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <parameters>
+        <parameter key="sylius.behat.page.shop.order.thank_you.class">Sylius\Behat\Page\Shop\Order\ThankYouPage</parameter>
+        <parameter key="sylius.behat.page.shop.order.show.class">Sylius\Behat\Page\Shop\Order\ShowPage</parameter>
+    </parameters>
+
+    <services>
+        <service id="sylius.behat.page.shop.order.thank_you" class="%sylius.behat.page.shop.order.thank_you.class%" parent="sylius.behat.symfony_page" scope="scenario" public="false" />
+        <service id="sylius.behat.page.shop.order.show" class="%sylius.behat.page.shop.order.show.class%" parent="sylius.behat.symfony_page" scope="scenario" public="false" />
+    </services>
+</container>
