+++ src/Sylius/Bundle/AdminBundle/Resources/config/routing/route.yml
@@ -20,7 +20,7 @@ sylius_admin_route_update:
             template: SyliusAdminBundle:Crud:update.html.twig
             vars:
                 templates:
-                    breadcrumb: SyliusAdminBundle:Route/Breadcrumb:_update.html.twig
+                    breadcrumb: @SyliusAdmin/Route/Update/_breadcrumb.html.twig
     requirements:
         id: ".+"
